requirements.txt
python-telegram-bot==20.7
groq==0.4.2
requests==2.32.3
flask==3.0.3
